refactor: extract composable createApp/startup for cloud overlay
Split index.ts into composable pieces that the cloud overlay imports:

- app.ts: createApp(container) — creates Hono app with all routes
- startup.ts: startConsumers(), startWorkers() — boot sequence
- server.ts: public API exporting all composable pieces
- index.ts: simplified to use the composable pieces directly

The cloud overlay (supaproxy-cloud) imports from server.ts:
  import { createContainer, createApp, ... } from 'supaproxy-server/server'

Added "exports" field to package.json for the subpath import.
All 149 tests pass.
